Current version, partially works
Works with occasional glitches. When the snake head and tail are on opposite sides of the board in the same row, it will sometimes crash. Also, the apple's spawn location occasionally crashed the game by making the program think that either the head of the snake or the apple is not in the list 'snake'.